integerating_Yolo_with_Postgresql_and_fastAPI/requirements.txt
fastapi
uvicorn
psycopg2-binary
ultralytics
opencv-python
numpy
